requirements.txt
pytest
allure-pytest
selene~=2.0.0rc9
allure-python-commons~=2.13.5